compilation warning fixed

+++ src/scag/pers/AbntAddr.hpp
@@ -275,7 +275,7 @@ public:
     }
     value.addr_content.type = (uint8_t)itype;
     value.addr_content.plan = (uint8_t)iplan;
-    value.addr_content.length = strlen(addr_value);
+    value.addr_content.length = static_cast<uint8_t>(strlen(addr_value));
     setValue(value.addr_content.length, addr_value);
     number = atoll(addr_value);
     if (number == 0) {
